// src/app/models/member.ts
export class Member {
    uuid: string;
    first_name: string;
    last_name: string;
    nick_name: string;
    membership_from_ts: string;
    membership_to_ts: string;
  }

  export class Manager {
    uuid: string;
    first_name: string;
    last_name: string;
    nick_name: string;
    manager_from_ts: string;
    manager_to_ts: string;
  }